+++ src/Disks/getOrCreateDiskFromAST.h
@@ -15,9 +15,4 @@ class ASTFunction;
  */
 std::string getOrCreateDiskFromDiskAST(const ASTFunction & function, ContextPtr context);
 
-/*
- * Is given ast has form of a disk(<disk_configuration>) function.
- */
-bool isDiskFunction(ASTPtr ast);
-
 }
