+++ src/Parsers/isDiskFunction.h
@@ -1,0 +1,9 @@
+#pragma once
+#include <Parsers/IAST_fwd.h>
+
+namespace DB
+{
+
+bool isDiskFunction(ASTPtr ast);
+
+}
